refactor: rename pre-diff to prediff
